+++ android/src/main/java/com/befovy/fijkplayer/RawMediaDataSource.java
@@ -79,4 +79,3 @@ class RawMediaDataSource implements IMediaDataSource {
     }
 
 }
-